Fix Travis-ci build (#4)
* Fix makefile
* Cast doubles to floats
* Upgrade travis-ci dist to bionic

+++ ArduMower/Gps.cpp
@@ -408,20 +408,20 @@ float Gps::distance_between(float lat1, float long1, float lat2, float long2)
   // Because Earth is no exact sphere, rounding errors may be up to 0.5%.
   // Courtesy of Maarten Lamers
   float delta = radians(long1 - long2);
-  float sdlong = sin(delta);
-  float cdlong = cos(delta);
+  float sdlong = static_cast<float>(sin(delta));
+  float cdlong = static_cast<float>(cos(delta));
   lat1 = radians(lat1);
   lat2 = radians(lat2);
-  float slat1 = sin(lat1);
-  float clat1 = cos(lat1);
-  float slat2 = sin(lat2);
-  float clat2 = cos(lat2);
+  float slat1 = static_cast<float>(sin(lat1));
+  float clat1 = static_cast<float>(cos(lat1));
+  float slat2 = static_cast<float>(sin(lat2));
+  float clat2 = static_cast<float>(cos(lat2));
   delta = (clat1 * slat2) - (slat1 * clat2 * cdlong);
   delta = sq(delta);
   delta += sq(clat2 * sdlong);
-  delta = sqrt(delta);
+  delta = static_cast<float>(sqrt(delta));
   float denom = (slat1 * slat2) + (clat1 * clat2 * cdlong);
-  delta = atan2(delta, denom);
+  delta = static_cast<float>(atan2(delta, denom));
   return delta * 6372795.0f;
 }
 
@@ -434,10 +434,10 @@ float Gps::course_to(float lat1, float long1, float lat2, float long2)
   float dlon = radians(long2 - long1);
   lat1 = radians(lat1);
   lat2 = radians(lat2);
-  float a1 = sin(dlon) * cos(lat2);
-  float a2 = sin(lat1) * cos(lat2) * cos(dlon);
-  a2 = cos(lat1) * sin(lat2) - a2;
-  a2 = atan2(a1, a2);
+  float a1 = static_cast<float>(sin(dlon) * cos(lat2));
+  float a2 = static_cast<float>(sin(lat1) * cos(lat2) * cos(dlon));
+  a2 = static_cast<float>(cos(lat1) * sin(lat2)) - a2;
+  a2 = static_cast<float>(atan2(a1, a2));
   if (a2 < 0.0)
   {
     a2 += TWO_PI;
